Custom decode errors for unhandled error types (#30)
Currently terraform is failing with the following error response:
```
Error: json: cannot unmarshal string into Go value of type sentry.APIError
```
Believe this is due to a rate limit error. To support different error
types this decodes the value to an interface which is typed switch for
pretty printing the error response.