fix: Update assertPostJSONValue to raise error on failing assetion
This commit is updating the assertPostJSONValue function in the
senty_test package to use `EqualValues` instead of `ObjectsAreEqualValues`.

The `ObjectsAreEqualValues` is returning a boolean representing if the objects
are equals or not, the issue is this value is never read during the
test. Switching to to `EqualValues` and passing the testing.T will raise
an error if one of the test is failing when checking the post JSON
value.

I updated the failing test accordingly.